+++ dds/idl/dds_visitor.cpp
@@ -243,9 +243,10 @@ dds_visitor::visit_structure(AST_Structure* node)
   if (idl_global->is_dcps_type(node->name())) {
     if (be_global->warn_about_dcps_data_type()) {
       idl_global->err()->misc_warning("\n"
-        "  DCPS_DATA_TYPE and DCPS_DATA_KEY pragma statements are deprecated; please\n"
-        "  use @topic, @key, @nested, and @default_nested instead. See the OpenDDS\n"
-        "  Developer's Guide for more information.", node);
+        "  DCPS_DATA_TYPE and DCPS_DATA_KEY pragma statements are deprecated; please use\n"
+        "  topic annotations instead.\n"
+        "  See docs/migrating_to_topic_annotations.md in the OpenDDS source code for\n"
+        "  more information.", node);
     }
   }